requirements.txt
PyQt5 >= 5.15
pyyaml >= 6.0
rich >= 12.4
executor >= 23.2
ipykernel >= 6.15
ipython >= 8
jupyter-client == 6.1.12  # higher versions will raise a TypeError on zmq poll ...
jupyter-core
qtconsole
pyzmq==23.2.0

pytest
pytest-order
